polybar: implement save to spotify library action
A python server is kept in an endless loop, outputting the currently
played song. Due to using spotipy, the server is able to connect to the
spotify web API with my user credentials, which are rightfully not in
this repository. This way, I can issue actions directly to the Spotify
service.

The spotipyClient script also formats the output for polybar,
underlining the song in orange (the current highlight color) if the song
is already saved to the library.
Experiments with different methods to highlight if the song has been
saved to library already have been done. A heart was proven to
disruptive. Changing the divider from '-' to '+' was too small of a
change.
The underline as highlight also fits quite nice into the current bar
design.

Currently, right-clicking the song display in polybar saves the
current song to the user's library.