compton: remove/change deprecated settings to new versions (6.2)
`no-{dock,dnd}-shadow` are deprecated and instead the option shadow of the
wintype dock/dnd should be used.
`menu-opacity` is deprecated and instead the option opacity of wintype
{popup,dropdown}-menu should be used.
vsync is now a boolean, so set it false.
glx-swap-method is deprecated and should be replaced by use-damage=true.
clear-shadow is now always enabled.
paint-on-overlay is always enabled if possible.
alpha-step is no longer required as compton now tries to make use of all
alpha values.