television: initial commit of xrandr output script and hotplugging udev rule
Due to a new, longer HDMI cable, I can finally use my television as a
fake smart TV and enjoy Netflix etc on a bigger screen.
The hotplugging rule for udev helps to autoconfigure it when I plug it
in.
The script just enables me to manually enable/disable the output.

Source:
https://frdmtoplay.com/i3-udev-xrandr-hotplugging-output-switching/