redshift: initial commmit of config and notification script
The notification script is more or less useless, but it helps to know
that Redshift has started shifting the colour instead of thinking I have
been in front of the PC too long.

The config is more or less standard, but it is focused on the second
output of xrandr (that's what the crtc line is for), so I can still use
my main monitor/the television for media or games.