locale: explicitly set all language settings provided by locale
This helped me to eliminate an issue with Steam which disallowed me from
using the keyboard in the chat.
Normally, this would/is handled by KDE's systemsettings5, but somehow
it did not work for Steam.

I also needed to reinstall libx11, so that X11 is capable of handling
all locales.
Not sure what went wrong here.
Maybe the one time I fiddled with pacman's option to exclude locale
files broke something.